Inventory 1.0 Finished (with a little bug)
If the player get into a close enough distance (InnerCircle) to the Item (BaseItem), the player can press E to collect the item. By pressing E, the object is destroyed and the item will be registered to the inventory as a 2D slot with the item's 2D image. To check the inventory, press I. The Item is stackable if they are exactly the same object (===). There is a bug when you collect 2 and more items.

“Attempted to access BP_Item_C_1 … but it’s not valid (pending kill/garbage)”

This could happens because DestroyActor on the item, but player still holds FocusedItem and tries to call the interface again (or the same press fires twice).

A BaseItem has 2 circles, outer circle (with outer icon) and inner circle (with inner icon). The outer circle was for signaling that the object is inspectable, and the inner circle is for signaling that the object is now interactable, but somehow they all disappear. For inner circle, it is because the mesh is too big and cover the icons, but the disappear of the outer circle could be a bug.